FINALLY DONE

+++ fib.c
@@ -3,7 +3,7 @@
 #include <string.h>
 
 // Fibonacci Functions
-unsigned long long recursive(int n)
+int recursive(int n)
 {
    if (n == 1)
       return 0;
@@ -12,22 +12,21 @@ unsigned long long recursive(int n)
    return recursive(n-1) + recursive(n-2); 
 }
 
-unsigned long long iterative(int n)
+int iterative(int n)
 {
-    if (n == 1)
+   if (n == 1)
       return 0;
    if (n == 2)
       return 1;
    
-   unsigned long long a = 0;
-   unsigned long long b = 1;
-   unsigned long long next;
+   int a = 0;
+   int b = 1;
+   int next;
 
-   //Fibonacci number #1 is 0 so start at 3 instead of 2.
+   //Fibonacci number #1 is 0 so start at 3 instead of 2. F(1)= 0, F(2) = 1
    for(int i = 3; i <= n; ++i)
    {
       next = a + b;
-      //printf("Step %d: a = %llu, b = %llu, next = %llu\n", i, a, b, next);
       a = b;
       b = next;
    }
@@ -38,7 +37,6 @@ int main(int argc, char *argv[])
 {
 
    int number = atoi(argv[1]);
-   //char info = argv[2][0]; // for "r" or "i"
    FILE *file = fopen(argv[3], "r");
 
    int fileNum;
@@ -47,9 +45,8 @@ int main(int argc, char *argv[])
    fclose(file);
 
    int N = number + fileNum;
-   unsigned long long fibNumber;
+   int fibNumber;
 
-   //printf("N = %d", N);
    if(strcmp(argv[2],"r") == 0) //recursive 
    {
       fibNumber = recursive(N);
@@ -59,7 +56,7 @@ int main(int argc, char *argv[])
       fibNumber = iterative(N);
    }
 
-   printf("%llu\n", fibNumber);
+   printf("%d\n", fibNumber);
    
 
 return 0;
